requirements.txt
cython
setuptools==59.5.*
opencv-python
opencv-contrib-python
prettytable
tqdm
yacs
torchfile
plyfile
munkres
chumpy
scipy==1.13.0
trimesh
shyaml
PyYAML
numpy==1.22.4
numpy-quaternion
pygame
keyboard
transforms3d
bvhtoolbox
vedo
imgaug
tensorboard
smplx
pycocotools
plotly
open3d
addict
loguru
norfair
h5py
imageio-ffmpeg
pyrender
joblib
pandas
